Build err (missing core); then, TODO test 926
Things should be in place to start testing at least the uart of the 926
(should hold off on screen). Getting a link/load issue with -core, need
to track down. May be related to reverting to 0.9 (now on an Arch box).

Started to pull out the PrimeCell UART driver from the 926 to its own
file (though that's not compiling yet). Turns out the 1176 uses the same
PL011 peripheral as UART0 (and a mini UART as UART1), so having the same
interface there is optimal. Plan is to have the struct & trait & statics
in primecell_uart (probably to be renamed), and to have instances in
arm926ej_s and arm1176jzf_s. The mini UART on the RPi may or may not use
the PL011 implementation - depends on compatibility.

Note that while the 926's UARTs are documented in its chip ref, the
Raspberry Pi's are documented in the peripheral reference for the
BCM2835 instead.